requirements.txt
# HTTP and API clients
requests==2.32.3
aiohttp==3.11.10

# Configuration management
python-decouple==3.8

# Device integrations
roku==4.1.0
pysmartthings==0.7.7

# Fitness/Health integrations
garminconnect==0.2.26

# User input utilities
pwinput==1.0.3
readchar==4.2.0

# GitHub integration
PyGithub==2.6.0

# OpenAI integration
openai==1.59.5